# feat(promhandler): use optimized API for `Series` queries