--- a/dev/local/ch-bench-read/testdata/bench-series.yml
+++ b/dev/local/ch-bench-read/testdata/bench-series.yml
@@ -1,0 +1,6 @@
+start: "2024-01-06T13:10:00Z"
+end: "2024-01-06T14:02:45Z"
+step: 15
+series:
+  - title: All series
+    matchers: ['{job=~".+"}']