--- a/internal/chstorage/chsql/_golden/Test2.sql
+++ b/internal/chstorage/chsql/_golden/Test2.sql
@@ -1,1 +1,1 @@
-SELECT (span_id) FROM (SELECT (span_id),(timestamp) FROM spans WHERE (true) AND (duration > 3.14) AND (duration < 3.14)) LIMIT 1
+SELECT span_id FROM (SELECT span_id,timestamp FROM spans WHERE (true) AND ((duration > 3.14)) AND ((duration < 3.14))) LIMIT 1