--- a/internal/chstorage/chsql/_golden/Test3.sql
+++ b/internal/chstorage/chsql/_golden/Test3.sql
@@ -1,1 +1,1 @@
-SELECT (timestamp) FROM spans ORDER BY timestamp DESC
+SELECT timestamp FROM spans ORDER BY timestamp DESC