--- a/internal/chstorage/chsql/_golden/Test4.sql
+++ b/internal/chstorage/chsql/_golden/Test4.sql
@@ -1,1 +1,1 @@
-SELECT (timestamp) FROM spans ORDER BY timestamp ASC,duration DESC
+SELECT timestamp FROM spans ORDER BY timestamp ASC,duration DESC