--- a/internal/chstorage/chsql/_golden/Test7.sql
+++ b/internal/chstorage/chsql/_golden/Test7.sql
@@ -1,1 +1,1 @@
-SELECT (column) FROM spans
+SELECT column FROM spans