--- a/internal/chstorage/chsql/sugar_test.go
+++ b/internal/chstorage/chsql/sugar_test.go
@@ -16,9 +16,9 @@ func TestInTimeRange(t *testing.T) {
 		want   string
 	}{
 		{"timestamp", time.Time{}, time.Time{}, "true"},
-		{"timestamp", time.Unix(0, 1), time.Time{}, "toUnixTimestamp64Nano(timestamp) >= 1"},
-		{"timestamp", time.Time{}, time.Unix(0, 10), "toUnixTimestamp64Nano(timestamp) <= 10"},
-		{"timestamp", time.Unix(0, 1), time.Unix(0, 10), "toUnixTimestamp64Nano(timestamp) >= 1 AND toUnixTimestamp64Nano(timestamp) <= 10"},
+		{"timestamp", time.Unix(0, 1), time.Time{}, "(toUnixTimestamp64Nano(timestamp) >= 1)"},
+		{"timestamp", time.Time{}, time.Unix(0, 10), "(toUnixTimestamp64Nano(timestamp) <= 10)"},
+		{"timestamp", time.Unix(0, 1), time.Unix(0, 10), "((toUnixTimestamp64Nano(timestamp) >= 1) AND (toUnixTimestamp64Nano(timestamp) <= 10))"},
 	}
 	for i, tt := range tests {
 		tt := tt
@@ -44,7 +44,7 @@ func TestJoinAnd(t *testing.T) {
 				Ident("foo"),
 				Ident("bar"),
 			},
-			"foo AND bar",
+			"(foo AND bar)",
 		},
 		{
 			[]Expr{
@@ -52,7 +52,7 @@ func TestJoinAnd(t *testing.T) {
 				Ident("bar"),
 				Ident("baz"),
 			},
-			"foo AND bar AND baz",
+			"(foo AND bar AND baz)",
 		},
 	}
 	for i, tt := range tests {