--- a/internal/metricstorage/metricstorage.go
+++ b/internal/metricstorage/metricstorage.go
@@ -1,0 +1,14 @@
+// Package metricstorage defines some interfaces for metric storage.
+package metricstorage
+
+import (
+	"context"
+
+	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/storage"
+)
+
+// OptimizedSeriesQuerier defines API for optimal series querying.
+type OptimizedSeriesQuerier interface {
+	OnlySeries(ctx context.Context, sortSeries bool, startMs, endMs int64, matcherSets ...[]*labels.Matcher) storage.SeriesSet
+}