… internal/chstorage/chsql/_golden/Test1.sql