```
                  │     base      │                 PR                  │
                  │    sec/op     │    sec/op     vs base               │
PromQL/All_series   2889.6m ± ∞ ¹   654.9m ± ∞ ¹  -77.34% (p=0.008 n=5)

                  │     base      │                 PR                  │
                  │   ch-sec/op   │  ch-sec/op    vs base               │
PromQL/All_series   2710.7m ± ∞ ¹   632.4m ± ∞ ¹  -76.67% (p=0.008 n=5)

                  │      base       │                   PR                    │
                  │ ch-mem-bytes/op │ ch-mem-bytes/op  vs base                │
PromQL/All_series     188.3Mi ± ∞ ¹     450.1Mi ± ∞ ¹  +139.10% (p=0.008 n=5)

                  │       base       │                   PR                    │
                  │ ch-read-bytes/op │ ch-read-bytes/op  vs base               │
PromQL/All_series     124.65Mi ± ∞ ¹      73.46Mi ± ∞ ¹  -41.06% (p=0.008 n=5)

                  │      base       │                  PR                   │
                  │ ch-read-rows/op │ ch-read-rows/op  vs base              │
PromQL/All_series      5.959M ± ∞ ¹      5.951M ± ∞ ¹  -0.14% (p=0.008 n=5)
```